REFACTOR LoginForm to react-hook-form and new file
